requirements.txt
# Core runtime dependencies for the Auto Content Factory pipeline
torch>=2.0.0
transformers>=4.40.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=9.0.0
ffmpeg-python>=0.2.0
